Alteração da rota e criação da entity.
