+++ go.mod
@@ -2,4 +2,4 @@ module github.com/sk8sta13/client-server-api
 
 go 1.13
 
-require github.com/mattn/go-sqlite3 v2.0.3+incompatible // indirect
+require github.com/mattn/go-sqlite3 v2.0.3+incompatible
